Update gendist.sthlp
The catalog of generated variables has been elaborated to clarify differences between those variables in unstacked and stacked data.
QUESTION:  The p_ values are stated here to be missing when respondent self-placements are missing, but a user might want to determine the position of a party even for respondents who did not place themselves. Can that be changed?
NOTE: When I ran the StackMe version from github gendist saved only d_variables; no m_* no p_*, no gendist_mc, etc. Presumably this is a bug, perhaps introduced when we changed the behavior of missing(diff)?