genmeans.sthlp
This description calls for non-trivial changes to the corresponding ado file. Note that implementing the cweight procedure for generating means that are constant across respondents for a given stack in a given context will call on genplace to derive the placements concerned. WE PROBABLY NEED TO DISCUSS THIS.  COMMENTS INVITED.