Update gendist.ado
Replaced with new version of gendist which is a syntax processor that extracts multiple varlists that use piped-varlist syntax and calls gendistP (the new name for the program that used to be gendist) once for each varlist it finds. GendistP, in turn, has been revised to use Stata's statsby context instead of levelsof context (so making only one pass thru the data instead of one pass for each context) and yielding an almost ten-fold increase in processing speed (14 seconds for a gendist that previously took 2 min 12 seconds).