Update genstacks.ado
Reworked to abstract just the variables to be stacked and, for each context, reshape those variables to long format and save in a separate file for each context. Then append each file  and merge the results with the original file. This takes about 1 minute 30 seconds (of which 20 seconds or so is spent in merging the files) in place of more than an hour with the previous genstacks for the same file containing 190-some contexts that yields a stacked dataset of over a million cases. There were a lot of challenges where Stata did not behave as expected, for example in not properly restoring preserved data, not appending a list of files, and some other problems. But the result is heartening. It would run even faster using data frames, but I am impressed by how fast it runs without.