Update genyhats.ado
Need a new option to set a missing data code for the y_depvar, if the ":" syntax is used. With a y_indep one can stash the missing data code actually encountered for that var, but that cannot be done when creating a y_depvar, which might have several missing data codes with no way of choosing between them. (This is a particular problem with datasets that have numeric missing data codes, like the TEV).