Update gendist.ado
Previously saved wrong version. This is version 2.0 for stackMe. It still runs on Stata Version 9 but includes code to detect whether the actual system is 16 or above, in which case it uses frames instead of a tempfile. Both syntaxes have been tested separately but not yet with the choice made at run time.