Update stackme.sthlp
This updated help text for StackMe as a whole introduces the optional alternative(s) to standard Stata variable lists and reframes much of the content to take account of minor improvements to StackMe that become available with Version 2.0. It needs to be read in conjunction with revised help files for the various commands that make up the StackMe package. At present documentation is running a bit ahead of implementation because code must live up to expectations given by help text.