Create stackmeWrapper.ado
stackmeWrapper is the central motor of stackMe 2.0. It is called by each 10-line caller program, named for each stackMe commend (here referred to as `cmd') and calls each `cmd'P (the original stackMe command with P for Program added to its name). Its main function is to shape the data seen by each `cmd'P, often context/stack by context/stack and generally subset according to [if][in] expressions accompanying the original command. Critical aspects of that shaping proccess are described in initial comments.