Update stackmeWrapper.ado
Wrapper program for stackMe version 2.0 commands. It is called by a caller program named for each command to execute code that all commands have in common. It then calls the appropriate specific command (also named for the stackMe command concerned, but suffixed with P for Program) known in this file as `cmd'P. For genyhats and geniimpute `cmd'P is called repeatedly once for each combination of context and stack. Other commands are,  for now, called only once. Mark will experiment to see if multiple calles makes them faster.