Update geniimputeP.ado
This ado file contains two programs (geniimputeP and geniimpute_body) that do the actual work of imputing, for each group of varlists having the same options and for each combination of stacks and contextvars, the requested imputations. It is called by stackMeWrapper, which handles tasks that stackMe commands have in common.