Update genstacks.ado
Revised in light of major changes to genstacksP and stackmeWrapper instigated by the need to be able to drop empty stacks after reshaping. These changes led to context-by-context execution, which required much of the code involved with naming and labeling the reshaped data to be moved to this ado file which contains the caller program that initiates genstacks processing.