Update stackmeWrapper.ado
Revised version attempts a logical progression thru tasks to be performed, initially focused on producing lists of variables to be kept in the working data, starting with variable lists. then variables mentioned in 'if' and 'weight' expressions, then variables mentioned in options. Program is unchanged after call on 'cmd'P; 'tho I plan to change the way it records tempfiles to be appended, context by context (these could now get very cumbersome with hundreds of contexts).