Update geniimputeP.ado
Minor tweaks to handle revised weighting strategy.

THIS PROGRAM HAS BEEN UPLOADED DESPITE A BUG THAT
ends execution when geniimputeP calls geniiP_body. Syntax errors in calls are really hard to find and I did not have time to persevere before leaving for Thanksgiving week. Other bugs might be found when that one is fixed, but should be minor.