Update stackmeWrapper.ado
Many minor changes to accommodate the partially implemented structure in which a `cmd'O program is called for all contexts prior to calling `cmd'P for each context in turn. The objecting is mainly to reduce the overhead code within 'cmd'P programs – code that has to be executed for each context and stack. But also to provide a point at which opening code involving all contexts can be executed before the code that is restricted to one context at a time.