Update stackmeWrapper.ado
This version (9) of stackmeWrapper incorporates a minor structural revision to accommodate 'genplace', which needs to be able to pass the dataset twice, context by context, if 'call' is optioned. That revision made it possible to move preliminary code from several of 'cmd'P, ensuring that code was only processed once instead of for each context. These revisions called for minor changes to each of the 'cmd'P programs that were shorn of their introductory code blocks. Knock-on effects were inevitable and called for extensive debugging.