Update stackmeWrapper.ado
The relevant description is attached to the previous version. It mentions redesigned elements in wrapper to facilitate the new genplace. Those elements made it possible also to separate other 'cmd'P programs into two parts (cmd'O and cmd'P) so that less code had to be processed for each context. The wrapper evidently had to be changed to insert the extra calls. In addition, diagnostics were largely moved into the wrapper from cmd'P programs, thus reducing redundancy and increasing conformity.