Update stackmeWrapper.ado
As mentioned in the previous two updates to this program, it’s main differences from earlier versions arise from the redesign required for it to be able to support a stackMe command that was really different from those comprising Version 1 (of what was then PTVtools). The additional command, genplace, is heretofore unique in requiring two passes through the active dataset, so as to accommodate two-step placements and calls on user-written subprograms. The resulting required change to the basic design opened the way to enhanced efficiency for existing commands that could have their initial codeblocks hived of to the newly available ‘cmd’O (for open), additional to  the previously existing ‘cmd’P (for program). This could significantly reduce the code that previously had to be executed for each context. Most ‘cmd’P programs have been significantly accelerated by these means, but at the cost of seemingly endless knock-on complications that took months to debug.