Update stackme.ado
As mentioned in the previous commitment of this ado file, it came into being as a result of a decision to take control of the naming of files, contextvars and item names. If this package is to be used for teaching it will be important that it not burden student users with extraneous concerns. It is imperative that a stacked dataset retain the contextvars that governed the stacking process, that filenames have distinct prefixes according to whether they are unstacked, singly stacked or doubly stacked, and that variables linking stacks to batteries be managed consistently. So stackMe utility programs where introduced to manage the complications that arise when automatic processes go wrong. The expectation is they these will rarely be used, except for SMcontextvars that has to be run as the first stackMe command following the opening (‘use’ing) of a datafile that is to be processed by stackMe commands.