Update stackmeWrapper.ado
As mentioned in the previous two updates to this program, it’s main differences from earlier versions arise from the redesign required for it to be able to support a stackMe command that was really different from those comprising Version 1 (of what was then PTVtools). The additional command, genplace, is heretofore unique in requiring two passes through the active dataset, so as to accommodate two-step placements and calls on user-written subprograms. This required change to the basic design opened the way to enhanced efficiency for existing commands that could have their initial codeblocks hived of to the newly available ‘cmd’O (for open), additional to  the previously existing ‘cmd’P (for program). This could significantly reduce the code that previously had to be executed for each context. Most ‘cmd’P programs have been significantly accelerated by these means, but at the cost of seeingly endless knock-on complications that took months to debug.
     This ado file also contains numerous subprograms, many of them consisting of chunks of code taken from the wrapper in an effort to make the central codeblocks less numerous and shorter, hence easier to understand. An additional subprogram handles error exits in a standard fashion (implying considerable changes to code in the wrapper itself). The biggest innovation here was to capture all errors (not just expected errors) by surrounding the whole wrapper by capture braces, except for a final codenlock that processes the errors once captured. That codeblock also clears all globals before exiting, except for three needed for communication between the wrapper and its calling programs. Capturing all errors (including any syntax errors that remain in untried code) should ensure that a user will never be left looking at a working dataset consisting of a single context-worth of observations.  Caller programs each have their own set of capture braces with the same end in view.