Update gendummies.ado
In Version 2 this program is now a caller named for the stackMe command that invokes it. It sets up an ‘options mask’ specific to the command, with options unique to that command. It then invokes program ‘stackmeWrapper’ that performs the command preprocessing for all stackMe commands before calling the specific `cmd’O (for Open, if extant) and `cmd’P (for Program), command-specific programs that do the heavy lifting for each stackMe command. Alone among V2 stackMe commands, gendummies’ ‘cmd’P is only called once for all contexts (since there is no context-specific processing for this command). Also, post-procesing in this caller routine is very limited.