Update stackmeWrapper.ado
I am traveling for the next week and want to put this file somewhere safe. There have been a lot of changes involving the reporting of errors diagnosed by Stata, to avoid messy program termination and giving Beta testers helpful information to pass on to the programmer. These changes have already born fruit in diagnosing errors in gendummies, which is in process of being fixed and geniimpute which is on the verge of running successfully. Remaining cmd'P are unchanged since November but should be easily updated by copying new code at start and end of gendist.ado to the ado files for other commands.