Update geniimpute.ado
n Version 2 this program is now a caller named for the stackMe command that invokes it. It sets up an ‘options mask’ specific to that command, with options unique to that command (options common to all commands are added in the wrapper). It then invokes program ‘stackmeWrapper’ that performs the command preprocessing for all stackMe commands before calling the specific `cmd’O (for Open, if extant) and `cmd’P (for Program), command-specific programs that do the heavy lifting for each stackMe command. Code in ‘cmd'P programs  is largely unchanged from the code used in the PTVtools version of stackMe., except that these programs are generally called repeatedly, once for each context. After finishing work on data for each context `cmd’P  returns control to the wrapper, which records the output for that context in a separate file and, after all contexts have been processed, concatenates all those files before merging the results back into the original dataset. The wrapper then returns control to the calling program which picks up where `cmd’P left off by post-processing the newly created variables (renaming and labeling them, etc., as optioned) before terminating execution with the outcome file left in active memory, in expectation of additional processing by other stackMe commands. 
   LATEST COMMIT has minor tweaks to accommodate stackMe-wide error handling.