Update stackmeWrapper.ado
This commit implements a fairly major update to the code for wrapping up the updating of the stackMe datafile in terms of variable prefixes and labels to document new variables created by a stackMe command. This was done separately for each command, within the caller program for that command. New code in the final codeblocks of stackMe wrapper now replaces the separate codeblocks previously found in each caller program.