Update genstacksP.ado
Minor tweaks that take advantage of the re-write of closing codeblocks previously located in program gendummies, now in subprogram cleanup (called by stackmeWrapper before return to geniimpute). Program genstacksP was not expected to be affected but because of its central position in stackMe workflow, changes in other programs were able to bring to light shortcomings in genstacksP that have here been addressed.