Update gendist.ado
gendist.ado
Part of a comprehensive upload of all stackMe Version2 ado files, revised to December 30th 2025, preparatory to a beta release. This ado file contains the caller program for a stackMe command. Additional ado files providing additional code needed for the same command are to be found in ado files with names starting with this command name and suffixed with the letter O (for Open) – preliminary code to be executed while all data contexts are still in memory – and P (for Program) – the original program file from ptvTools (version 1 of stackMe), slightly amended to handle the fact that it is called repeatedly, once for each data context.
   Still need to check each command with each combination of options, which may reveal defects in the wrapper or (more likely) one or more of its subprograms (especially ‘cleanup’) and fix any resulting bugs.