Update geniimputeP.ado
geniimputeP.ado
One component of a comprehensive upload of all stackMe Version2 ado files, revised to December 30th 2025, preparatory to a beta release. This ado file contains the processing code for a stackMe command. Additional ado files providing additional code needed for the same command will be found in an ado file with name starting with this command name and suffixed with the letter O (for Open) – preliminary code to be executed while all data contexts are still in memory.
      Still need to check this command with each combination of options, which may reveal design or coding defects, and fix any resulting bugs.