Update genmeanstats.ado
cmdCaller.ado
Part of a comprehensive upload of all stackMe Version2 ado files, revised to December 30th 2025, preparatory to a beta release. This ado file contains the caller program for a stackMe command. Additional ado files providing additional code needed for the same command are to be found in an ado file with names starting with this command name and suffixed with the letter P (for Program) – the original program file from ptvTools (version 1 of stackMe), comprehensive amended to handle the fact that it is called repeatedly, once for each data context and to expand the number of bivariate statistics it can produce.
      Still need to check each command with each combination of options, which may reveal defects in design or coding, and fix any resulting bugs