Update genstacksO.ado
genstacksO.ado
Part of a comprehensive upload of all stackMe Version2 ado files, revised to December 30th 2025, preparatory to a beta release. This ado file contains the opening program for a stackMe command. Additional ado files providing additional code needed for the same command are to be found in ado files with names starting with this command name and suffixed with the letter P (for Program) – the original program file from ptvTools (version 1 of stackMe), slightly amended to handle stackMe extensions, especially the need to called separately for each context
   Still need to check each command with each combination of options, which may reveal defects in the wrapper or (more likely) one or more of its subprograms (especially ‘cleanup’) and fix any resulting bugs.