Update genyhatsP.ado
genyhatsP.ado
Part of a comprehensive upload of all stackMe Version2 ado files, revised to December 30th 2025, preparatory to a beta release. This ado file contains the processing code for a stackMe command. An additional program is contained in this ado file, 'predcent', which does the heavy lifting for the command.
      Still need to check this command with each combination of options, which may reveal design or coding defects, and fix any resulting bugs.